Use interned string for PathElement as std::string is self referential (Ubuntu 16.04)
