+++ sql/testing/sql_testing.mk
@@ -3,5 +3,6 @@
 # This file is part of MLDB. Copyright 2015 mldb.ai inc. All rights reserved.
 
 $(eval $(call test,path_test,sql_types,boost valgrind))
+$(eval $(call test,path_order_test,sql_types,boost))
 $(eval $(call test,path_benchmark,sql_types,boost))
 $(eval $(call test,eval_sql_test,sql_expression,boost))
